hypersyncdt/backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
sqlalchemy==2.0.27
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pydantic==2.6.1
numpy==1.26.3
pandas==2.2.0
scikit-learn==1.4.0
torch==2.2.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dateutil==2.8.2
joblib==1.3.2
mlflow==2.10.2
plotly==5.18.0
streamlit==1.31.1
pytest==8.0.0
pytest-asyncio==0.23.5
httpx==0.26.0
black==24.1.1
isort==5.13.2
flake8==7.0.0
